Extract executeFunctionCall.
